+++ CHANGELOG.md
@@ -37,6 +37,8 @@ All notable changes to this project will be documented in this file.
 - Semantics of the existing method ``hilbert_space::dim()`` has been changed:
   Now it returns the exact dimension of the Hilbert space, which is smaller than
   ``hilbert_space::vec_size()`` if the Hilbert space is sparse.
+- New method overload ``hilbert_space::dim(elementary_space<...> const& es)``
+  that returns dimension of a constituent elementary space.
 - Changed ``space_partition`` to store a constant reference to the Hilbert space
   being partitioned. This change is necessary to enable support for the sparse
   ``hilbert_space`` objects. ``space_partition`` is now templated on the Hilbert